move slackbot inside planobs
